refactor: reorganize the project structure
