--- src/components/Video/Video.module.scss
@@ -1,3 +1,0 @@
-.video {
-  max-width: 100%;
-}
