+++ src/containers/MainPage/CreateRoomView/CreateRoomView.module.scss
