+++ src/containers/MainPage/CreateRoomView/reducer.js
