+++ src/containers/MainPage/MainPage.module.scss
